pyproject.toml
[project]
name = "captains-log"
version = "0.1.0"
description = "Voice-based ship's log for Sailing Vessel Defiant"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "boto3>=1.34.0",
    "openai>=1.40.0",
    "dbos>=0.9.0",
    "pgvector>=0.2.5",
    "asyncpg>=0.29.0",
    "ffmpeg-python>=0.2.0",
    "fastapi-users[sqlalchemy]>=13.0.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx-oauth>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",  # parallel test execution
    "pytest-it>=0.1.4",     # BDD-style test decorations
    "vcrpy>=6.0.0",         # record API calls
    "httpx>=0.27.0",
    "moto[s3]>=5.0.0",      # mock AWS services
    "pytest-mock>=3.12.0",  # enhanced mocking
    "factory-boy>=3.3.0",   # test data factories
    "faker>=24.0.0",        # generate fake data
    "aiosqlite>=0.19.0",    # async SQLite for testing
    "aiohttp>=3.9.0",       # async HTTP client for workflows
    "playwright>=1.40.0",   # browser automation for E2E tests
    "pytest-playwright>=0.4.4",  # pytest plugin for playwright
    "prospector[with_everything]>=1.10.0",  # code quality analysis
    "black>=24.0.0",  # code formatter
]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic/versions
)/
'''